Update health monitor to use v2 storage
HAProxy health monitor agent was still using the legacy DataClient and
ZkManager to access zookeeper.  This patch updates the agent code so
that it accesses zookeeper using the new v2 backend storage engine.

This patch also includes a fix for a bug in which two health monitors
attached to a router deletes one of them mistakenly.

Change-Id: I6aa2bce8d996ec172d1577a3278bd922e57462dd
Signed-off-by: Ryu Ishimoto <ryu@midokura.com>
Fixes: MNA-777
Fixes: MNA-832