Update health monitor to use v2 storage
HAProxy health monitor agent was still using the legacy DataClient and
ZkManager to access zookeeper.  This patch updates the agent code so
that it accesses zookeeper using the new v2 backend storage engine.

Also, for leader election, the recipe provided by CuratorFramework is
now used.

This patch also includes a fix for a bug in which two health monitors
attached to a router deletes one of them mistakenly.

Finally, the health monitor unit tests that still use DataClient are
commented out temporarily.

Note that even after this patch, the health monitor agent still updates
the network configuration data directly as part of midolman, which goes
against the v2 architecture design.  This is a temporary solution to
get the failing gate tests to pass for L4LB/LBaaS.  For the final
solution, either the health monitor agent will be changed to only
update the network state data, or moved to the cluster altogether.

Change-Id: I6aa2bce8d996ec172d1577a3278bd922e57462dd
Signed-off-by: Ryu Ishimoto <ryu@midokura.com>
Fixes: MNA-777
Fixes: MNA-832