Squash api and client into cluster, fix cobertura
This patch focuses on getting coverage working on all the new code in
midonet-cluster.

The midonet-api module contained only API tests for code that is in
midonet-cluster, but Coverage reports don't work cross-modules, so we
move the tests along the code into midonet-cluster.

Because the tests depend on midonet-client, and this module had
dependencies on cluster, so this created a circular dep that we resolve
by moving midonet-client into cluster.

Additionally, we remove here nested objects from scala code, since these
break cobertura reports, making minor warning and style cleanup in
affected areas.

Change-Id: I799493387ad46ce3387ed735e8c416ff460c003b
Signed-off-by: Galo Navarro <galo@midokura.com>