Agent: Review recommend heap sizes
As of now, flow state still requires us to have a pretty sizeable
heap, otherwise our latency suffers dramatically due to constant
pauses.

For computes, a heap of 2GB is adequate as tail latencies are okay
under 10kpps. For large computes, a heap of 4GB allows us to handle
20kpps of flow state generating packets.

For gateways, we can use 8GB to get better latencies. For
large gateways, we suggest 10GB, as it improves latency a little bit
as well as throughput. Up to 32GB (when we won't be able to use
compressed OOPS unless objects are aligned at 16 byte boundaries), we
expect to see small increases in latency and throughput as we keep
increasing heap size. However, these are not big enough for us to
recommend doing so.

Change-Id: I4ba87ed36ad326ea84abb44122c19a9314c348e0
Signed-off-by: Duarte Nunes <duarte@midokura.com>