Update patch set 1
Patch Set 1:

Well, not a load maybe (ofc it depends on the load). The curious thing is that every time the mapping is being removed, it's never found and the error is shown. I can fix the commit message removing "a load of", but the log shouldn't be an error by looking at the code anyway.

Patch-set: 1
Reviewer: Xavi León <1003318@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0