Update patch set 12
Patch Set 12: Code-Review-1

I'm personally not convinced that we need to add subjects for networks, as noted in #10.  If the goal is asynchrony, we have Futures which seem a much more adequate tool.  Since I'll have to pick this one up to join to the vxgw service from the other branch, I'll see how the alternative looks.

Patch-set: 12
Label: Code-Review=-1