Agent: Prefer IP total length value to buffer size
When deserializing an IP packet, we would prefer the buffer's
remaining size over the one specified in the total length header.

One particular example of where this is wrong is when we receive a
packet that is Ethernet padded to 60 bytes. We would override the
total length field, thus making the pad bytes part of the upper
protocol's payload. By replacing the total length we would also get a
bad checksum, since OVS selectively recomputes the checksum field by
field (that is, only for the fields changed with the SetKey flow
action) and doesn't totally recompute it; it would never take the new
total length header value into account.

Ref MN-3900

Change-Id: Ieb918a390e4dc30101408d723673d202187b0ce5
Signed-off-by: Duarte Nunes <duarte@midokura.com>