MDTS: add docker logs upon test failure
Some configuration and setup is done in the run-[service].sh script
(calling mn-conf, setting up configuration files to connecto to ZK and
cass, etc.). This patch adds the output of docker logs container_name to
the logs tarball for aid in debugging. This also removes the cluster
upstart log file from the report as it's redundant with the
midonet-cluster log.

Change-Id: Idc4afe835613e2a6b77b1fc88d3dc6a1069e9e9d
Signed-off-by: Xavi León <xavi.leon@midokura.com>