Add ServicePort
Part of service insertion. A ServicePort is a BridgePort with failOpen
functionality. If failOpen is enabled, when a packet egresses the
ServicePort while it is inactive or in adminState down, the packet will
behave as if it has ingressed the port (i.e. it will look like it has
been bounced back from the service).

ServicePort only accepts packets which have been redirected. Other
packets are dropped. It also ignore it's outfilters, since they are not
used for redirected packet.

Change-Id: I9e26b35c52de6b3fd8d23050f1b60909f4fe7d59
Signed-off-by: Ivan Kelly <ivan@midokura.com>