MDTS: retry connections to fix l4lb flakyness
Netcat clients sometimes fail to connect to the backend server under
load and returns an empty string as a result. This counts as a different
host in the assert and fails, as the result contains one more host than
expected. This patch address this by retrying a few times when the
return code of netcat is not zero.

Ref: MNA-833

Change-Id: If8fb36374018e5d9ed1031dd1e89e696491fef6c
Signed-off-by: Xavi León <xavi.leon@midokura.com>