Update devmido to use installShadowApp
This patch contains devmido adjustment to two recent changes in
midonet:

Commit 227d18cc5bb40bc2b33ccd1f4d98f610ea6ea24b moved mn-conf from
midolman to midonet-tools sub-project.

Commit a46e5ec8f2837965540ac8a1b6e35cb3ba728489 introduced a new way
to produce midolman and midonet-tools jars, which bundles all of their
dependent jars into one jar.

Update devmido so that mm-ctl, mm-dpctl and mn-conf proxy scripts point
to these new jars instead of creating a directory of the dependent
jars, for midolman and midonet-tools subprojects.

Fixes: MNA-875
Signed-off-by: Ryu Ishimoto <ryu@midokura.com>

Change-Id: I9699584fc8d2415b7530baa4abc9a68bb985bd5a