MDTS: remove unnecessary delay after agent restart
This patch removes a 30s delay after agent restart which is now
unnecessary as we already wait for the agent and ports to be up
(checking the midonet-api). This will solve the issue with conntrack
entries for rev-nat being removed too soon before we actually send the
return flow in the tests.

Ref.: MNA-803
Change-Id: If90f532d22ecf34c60c92e6d69bf2707b65ff895
Signed-off-by: Xavi León <xavi.leon@midokura.com>