Zoom ACL: State storage becomes host-aware
This change modifies the StateStorage interface to introduce
methods that will allow state read operations to distinguish
between different hosts. This will allow the separation of state
data by host and implementation of ACLs.

Currently, only read methods allow the selection of the host
identifier from which the state key-values are read. Write methods
will always use the current host identifier.

Only the method prototypes are defined in this patch, with the
implementation to follow in a subsequent patch. To provide
compatibility with the current code, the MidonetBackendService uses
`any` as host identifier, until support for multiple hosts is
implemented.

Ref: MNA-806

Change-Id: Ia46d357db4283d3ca371bb67e04b3504046f425f
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>