Zoom ACL: Multi-host support in the ZK state storage
This adds the implementation for multiple hosts to the ZooKeeper
state storage (ZookeeperObjectState) and corresponding unit tests.
With this change the state ZOOM clients can specify a host
identifier when creating the client instance. This client
identifier creates a separate namespace in the ZooKeeper tree for
the stata data of each hosts.

Read methods (get and observables) provide the capability to read
the state of different hosts. However, the write methods only
allow to write to the namespace of the host for which the client
instance was created.

As a difference from the previous behavior, because each host
namespace includes a private copy of the state key-value mappings,
the state keys are no longer created with the creation of the
corresponding model object. Instead they are created on demand,
whenever a host writes to the key. Additional changes ensure that
the object for which the state is added exists, and read methods
can accommodate the situation where the state keys are missing.

Also, the deletion of a model object no longer deletes the
corresponding state keys from each host's state namespace. Instead
a lazy cleanup mechanism will be provided in a different patch.

Ref: MNA-806

Change-Id: I84917252d80ad2126fbee3e178db2770382e23d3
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>