Bridge mapper ignores VXLAN ports
The bridge mapper should ignore the local ports of a bridge which
are of VXLAN type. Currently, it doesn't do so, assuming that all
ports of a bridge are bridge ports. Because of this, whenever a
VXLAN port is added to a bridge, the bridge mapper will fail with
a ClassCastException, preventing the virtual topology to load
any bridge that has one or more VXLAN ports.

This change validates that a bridge's local port is a bridge port
and ignores it otherwise.

Fixes: MNA-883

Change-Id: I29ce6aeac662232ed541e6e444b94885a31ef05c
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>