API: VTEP create should use HostResource
The VTEP create API fails when parsing the strings because it's
retrieving the Host DTO directly without going through the Host API
which composes the addresses from the host state.  This makes it choke
and NPE when doing IP verifications.

This patch fixes this problem, having VTEP go via HostResource.list.  It
also fixes HostResource, which was serializing the addresses incorrectly
from InetAddress so the format broke deserialization from IPv4Addr.

Ref.: MNA-890
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: Icf224322a43766ac6fa07347230d212ea0229379