[WIP] MDTS: add multiple session bgp tests
This patch adds multiple sessions tests to mdts. The default flavour now
spawns three quaggas, two for peering with midolman and a third to act
as an external host on the internet. The test itself creates two
separate sessions with two separate quagga daemons on the same interface
reachable on the same subnet and checks that packets are not lost under
different failure scenarios.

Change-Id: I890332223691c404282830d1ef5953982b4adcd5
Signed-off-by: Xavi León <xavi.leon@midokura.com>