Fix race publishing initial flooding proxy
The old code had a race between two processes: the first one is orchestrating
the initial connection, part of which is hooking the processor exposed by the
OVSDB client that applies Mac Remote writes to the VTEP. The second is a
general purpose connection watcher triggered every time that the connection is
ready (this can happen in the first one, but also after transient
reconnections). The watcher syncs the NSDB into the OSVDB, part of which
involves loading newly bound networks and for those, writing the logical
switches to the OVSDB and feeding the flooding proxy.

A race may occur when:

- The service starts and triggers a connection to the VTEP, sleeps
- The vtep reports a connection establishment, sleeps
- The general purpose watcher triggers a sync. Because so far, the hook to
  write mac remotes is not set up yet (this is done by the initial connection
  sync process), the flooding proxy announcements are dropped on the floor
- The initial connection wakes up and triggers a sync. It hooks the mac-remote
  pipe to the VTEP, and then syncs the networks bound to the vtep. Because this
  was just done in the previous step, it doesn't find any new networks to bind,
  we assume that the logical switches are created and configured (which is
  correct) and that the flooding proxies were published (which is not.

We now consider this case, and make the initial connection always publish the
FP to all networks after the pipe to write mac locations is set up.  We also
modify the watchers so VTEP is considered connected when the state goes to
READY, not CONNECTED.  (The former means not only that we have a connection,
but also that we have loaded all contents of the DB in the local cache)

Ref.: MNA-897
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: I8fa8429debccc3a2e3d3f1058f48a72a525da345