Don't set network route for gateway port
For FIP to FIP connection to work, the network route on the tenant
router must be removed.  The reason is that, with the route, FIP
traffic ingressing into the external network from the tenant router
immedaitely egresses out into the tenant router, and the midolman agent
does not support this case.  To make inter-FIP connectivity work, let
all FIP traffic leave the external network and get handled by the
uplink router.

Change-Id: I0a238338b77d3516d9d5f40259b5c2201c68d4e3
Signed-off-by: Ryu Ishimoto <ryu@midokura.com>
Fixes: MNA-903