midonet-cli lists VXLAN port correctly
This changes the midonet-cli and the cluster REST API to ensure the
correct listing of the VXLAN ports. The following changes are made:

- The Bridge DTO includes a unique URL for a bridge's VXLAN ports.
These are backed by a `BridgeVxlanPortResource` which returns the
set of VXLAN ports for that bridge.

- The midonet-cli only uses the Bridge `portIds` field when listing
the ports of a bridge. This is because these identifiers already
include the VXLAN ports.

- The API BridgePortResource supports both BridgePort and VxlanPort.
The allows the resource to return a list with the correct port
type when listing the ports of a bridge.

- The midonet-cli allows the Vtep object to have an id, which ensures
that the VTEP objects are properly aliased. It also changes
the `management-ip` field of a VXLAN port to `vtep`, which is now
included as a VtepRef. Hence, when listing the VXLAN ports for a
bridge, they will include a reference to the appropriate VTEP as:

  port port0 device bridge0 state up plugged no vtep vtep

Also, listing the VTEP includes the corresponding alias as:

  vtep vtep0 name vtep0 description  management-ip 127.0.0.1 ...

Ref: MNA-881

Change-Id: Ie9f2f5cfe676c1adc99ee385270a9589a3ac2e31
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>