VXGW: Fix OVSDB setBindings
There exists a bug in the OvsdbVtepData.setBindings which replaces
the bindings for all the logical switches bound to a set of ports,
rather than just the specified one. In addition, the current
documented behavior of this method is that only the bindings for
the physical ports specified in the agument list are modified,
regardless of whether other ports have been previously bound to the
same logical switch.

With the new behavior, the `setBindings` method replaces the
bindings for all physical ports previously bound to the given
logical switch with the one in the specified list.

Ref: MNA-901

Change-Id: Ib4935faee903618ebfe544f9fc6e381dec6c0c94
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>