VXGW: Prevent duplicate MAC entries
This enhances the OVSDB data client to handle the case where an
identical MAC location already exists in the OVSDB database. The
patch also cleans-up the code, and currently ensures that a MAC
can only correspond to one physical locator for a logical switch.
If the same MAC with a different IP address and locator is added,
then the old MAC entry will be deleted.

Ref: MNA-895

Change-Id: I0984557d336076364bf321d7c2a6e94625880cd3
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>