VXGW: The VXLAN port mapper exposes the current VTEPs
In MidoNet version 1, the VTEP tunnel IP and tunnel zone identifier
were stored in the VXLAN port. This information was removed in the
current architecture, where the VXLAN port contains only the VTEP
identifier. Therefore, when forwarding packets via a VXLAN port the
agent must load both the port and the VTEP in order to determine the
tunnel end-point.

With this change, we expose the VTEP tunnel information (tunnel IP
and tunnel-zone identifier) to the simulation via the VXLAN port
mapping service. This is a simpler alternative to having the VTEP
data loaded through the port mapper for a VXLAN port, since the VXLAN
mapping service already monitors all VTEPs. Hence, the simulation
may simply query the tunnel endpoint data using the VTEP identifier.

Ref: MNA-907

Change-Id: Ia1a7f8bf3379585fe39d3407f960bc09e3a7ba8a
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>