Update Guava to 18.0
Despite a declared dependency on Guava 16.0, the command line build
has been using 18.0 for quite some time, presumably due to a
dependency pulling it in. However, IDEA has continued to use 16.0. A
recent change added a call to MoreExecutors.directExecutor(), which
was added in 18.0, breaking the IDEA build.

This change fixes the IDEA build, and replaces the remaining calls to
MoreExecutors.sameThreadExecutor() to directExecutor().

Change-Id: Ibcc58a343fd9371b30fb8cadb42b6bb22f328e67
Signed-off-by: Brandon Berg <bberg@midokura.com>