Update patch set 3
Patch Set 3:

Actually. Sending a packet to warm up the topology would need its own timeout too, so in order to not be flaky it would need a long enough timeout, which would, for all intents and purposes, be the same thing as just increasing the timeout in the packet the test is already sending.

I think this test is specially prone to this because it needs to load more topology (an extra mirror, bridge and port) than a normal router test.

I upped the timeout to 15 secs per packet.

Patch-set: 3
Reviewer: Guillermo Ontañón <1002763@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0