Delete port chains on conversion from VIF to RIF
Creating a Neutron RouterInterface using a VIF port converts that VIF
port to a router interface port. When doing this,
RouterInterfaceTranslator was failing to create operations to delete
the VIF port's chains, which are not needed for RIF ports. This patch
fixes that, and also factors the router port creation out of
RouterInterfaceTranslator.translateCreate(), which was getting too
long.

Change-Id: I6d00b5d05b886fde429bce4f5cab36b5242e40c8
Signed-off-by: Brandon Berg <bberg@midokura.com>
Fixes: MNA-922
(cherry picked from commit 3165f7510705b5867a52d4d65e5de61322fee1de)