Agent: Be kinder to ZK when allocating Nat blocks
When allocating a NAT block, we asynchronously fetch the ZK nodes
corresponding to the blocks in the specified range. If these blocks
have not yet been created, all these gets will fail, and we will
create the directory structure. However, all operations then try to
create each node and its parents individually, which may lead to
multiple, failed attempts to create the nat/device/ip path. With this
patch, if we fail to fetch a node, we transactionally create all of
the directory structure, synchronously, before retrying the block
reservation.

Change-Id: Ieafd912dd1710891250b667a66885b19a34c51b2
Signed-off-by: Duarte Nunes <duarte@midokura.com>