Agent: Adaptively increase buffer size
This patch catches BufferOverflowExceptions in the
DisruptorDatapathChannel,
so as to adaptively increase the buffer size. On each exception, we
double
the buffer size, which starts at 64k and grows to a maximum of 4MB.

A BufferOverflowException may happen if a packet is received from a
network with an MTU over the buffer size or if we're executing a
packet or creating a flow and the size of the serialized flow actions
exceed the buffer size.

Change-Id: Ib79e3cef1db7175a3d1ced97739d0262f6440b1f
Ref: MN-4004
Signed-off-by: Duarte Nunes <duarte@midokura.com>