BGP: change bgpd configuration, add multisession tests
  * adds the multipath-relax option to let the midonet vrouter learn
  multiple paths to the same prefix from different peers.
  * disables the announcement of routes learned from other ASes

This patch also adds multiple sessions tests to mdts. The default flavour
now spawns three quaggas, two for peering with midolman and a third to act
as an external host on the internet. The test itself creates two
separate sessions with two separate quagga daemons on the same interface
reachable on the same subnet and checks that packets are not lost under
different failure scenarios.

Fixes: MNA-925
Signed-off-by: Xavi León <xavi.leon@midokura.com>
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>
Change-Id: I890332223691c404282830d1ef5953982b4adcdc