Update patch set 1
Patch Set 1:

Quoting Guillermo:

"IMHO the FPH rewrite is unwarranted and, should this code be needed in v5.0 (alex please confirm), or v1.9, it should stay out of those branches since they are code frozen."

My reply:

Actually reading the code, you can see it's not a re-write, it was me adding the ability to filter by host ID, fixing some stuff, and publishing the flooding proxies as an array list. One thing to keep in mind is that there were a lot of changes between the first patch and the final one, so when i rebased things might not have been applied as cleanly as possible (I know some stuff got reordered). Finally, I even added some coverage to the FloodingProxyHerald :-)

Patch-set: 1
Label: Code-Review=0