MDTS: l4lb tests for sender in the same pool subnet
This patch adds a couple of bindings and generalize the tests to include
the scenario in which the sender and pool members are on the same
subnet. For this, we need to add snat and rev_snat rules to the outbound
and inbound filter in the router (this is done automatically in the
neutron translators).

Ref.: MNA-926

Change-Id: I543f85467a386d9a74e36fb7b17e84f5dccd49a4
Signed-off-by: Xavi León <xavi.leon@midokura.com>