Agent: Iteratively traverse FloodActions
A FloodAction is a binary tree of arbitrary depth. In the particular
case of the bridge flood action, the FloodAction will resemble the
degenerate case of a tree, where all subtrees are on the right hand
side. Since we were traversing FloodActions recursively, bridges with
lots of ports (> 1000) could trigger a StackOverflowException.

This patch ensures that we do the traversal iteratively using a
stashed ArrayList as a stack to hold pending simulation actions.

Ref: MN-4014
Change-Id: Idf5c590304d5cdf1e6ac6ca2772da09e7f002721
Signed-off-by: Duarte Nunes <duarte@midokura.com>