L4LB: Health monitor supports multiple pools
This changes the ZOOM topology model to allow an L4LB pool health
monitor to support multiple pools, as oposed to only one. This
bringes the following changes.

(1) The Protocol Buffer and ZOOM bindings change and are incompatible
with previous versions.

(2) The REST API is updated to support multiple pool operations (e.g.
listing the pools of a health-monitor). The URI for a health monitor
pool has also changed to:

  /health_monitor/:health_monitor_id/pools

(3) The C3PO health monitor translator allows the creation of a
health monitor with zero or more pools. This changes the previous
behavior which allowed only one pool. The update operation has
not changes, and still prevents any modification to the set of
pools.

Ref: MNA-959

Change-Id: I8e5a366cb5633c7d8a7986a547c1c06ff5575371
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>