MDTS: l4lb add health monitoring tests and haproxy resilience tests
This patch adds tests for haproxy fault tolerance in which the midolman
in charge of hosting the health monitoring resources is rebooted and another
midolman is elected as the leader to hold the instances. It also checks
that all pools has the corresponding haproxy instance and that they get
cleaned when the topology is removed. This patch also adds a test to
check health monitors in which we alternatively fail backends and check
results.

TODO: this patch contains bits of the virtual topology commented and the
health monitor tests disabled so they can be enabled on subsequent
patches.

Change-Id: I27e1494b48fb629d5a24db08cfba943265c3abdf
Signed-off-by: Xavi León <xavi.leon@midokura.com>