Agent: Create PacketsEntrypoint actor before DpC
This patch ensures we create the PacketsEntrypoint is created before
the DatapathController, since the latter relies on the former to get
the total number of worker threads when creating tunnel ports.

It also ensures we retry to get the workers from the PacketsEntrypoint
in case a previous attempt fails, possibly because of a timeout.

Change-Id: Ib91272b2cafa28f5c44037534dbbecd117c14ce7
Signed-off-by: Duarte Nunes <duarte@midokura.com>