Auth: Fixes/improves tenant listing
Previously the API tenant resource returned a list of
o.m.c.auth.Tenant objects rather than o.m.c.rest_api.models.Tenant.
This patch addresses this, and include two additional fields that
are available from Keystone: the tenant description and whether
the tenant is enabled.

It also makes the necessary changes in the midonet-cli, to ensure
that all tenant fields are displayed. In addition, it returns
an error if the client attempts to write to the tenant resource.

The tenant media type has been bumped to v2.

Change-Id: I1ff40b4a53f491cd32f3603ea61605c3a328ca86
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>