Use 0.0.0.0 for source addr if none configured
When Midonet replies to DHCP requests, it first checks to make sure that
a server address is set so that the reply has a proper source address.

However, because our DHCP server works off of a bridge, a router may
not be present to act as the source address. In this situation, we still
need to send a DHCP reply.

This patch sets the DHCP source address to 0.0.0.0 in this scenario.

Fixes MNA-968

Change-Id: Ibcc2437fab81a9ce9c772b12f0639ebae2705feb