BGP: Publish bgpd status
This modifies the MidoNet backend to expose the status of the BGP
sessions as a per-port state value. This information is further
exposed via the API as a `bgpStatus` value for the RouterPort
objects.

The patch also includes the new BGP status field in the midonet-cli
and removes the legacy BGP classes.

Ref: MNA-971

Change-Id: I7007d7a46cb6fb8e707d1ad1746de9a453b45489
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
(cherry picked from commit dc3fd1a3df242726d8b3f6235df88bfd5b1bee9b)