Metering: count the 1st packet of each flow
Make the agent count usersspace-processed packets towards
metering. Typically this are the 1st packet of each flow,
but they can also be packets generated by virtual devices or
flows that are not implemented in the kernel.

Adds a length() method to the IPacket hierarchy and a
recordPacket() method to the meter registry interface.

Change-Id: I049f3fa8072a95fa5d9bcfaad92e25d2a711162a
Fixes: MN-3947
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>