Agent restarts immediately after failure
This changes the write policy for the host state keys (AliveKey and
HostKey) from SingleFirstWriteWins to SingleLastWriteWins. This
allows the host service of an agent to claim ownership of the state
keys immediately after a failure (i.e. event if the keys exist),
rather than waiting for ZooKeeper to delete the ephemeral nodes of
the previous state.

In order for a new agent instance to claim ownership over existing
state keys written by a previous session, the following conditions
must be met:
(i) the hostname should not have changed, and;
(ii) if a previous state entry exists for the HostKey, the list of
network interfaces must be the same.

If the state keys exist, but the above conditions are not met, the
host service will revert to the previous behavior:
(i) waiting for the state keys to be deleted from ZooKeeper, or;
(ii) shutting down if the keys are not deleted within the configured
timeout interval.

Fixes: MNA-988

Change-Id: Id627c8096d990fb3f8a1ed11c4e10cb35c714702
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>