Agent: Propagate resurrected MAC-port pair to ZK
When an agent detects a MAC has moved, it doesn't update the local
TimedExpirationMap used to expire MAC learning entries. This maps acts
as a sort of first level cache to the underlying ReplicatedMap.

If the MAC associated with a port moves from MAC X to MAC Y, the local
flows will be invalidated and the entry will move to the expiration
queue of the TimedExpirationMap, where it will stay for 30 seconds. If
in this time period a packet arrives moving the MAC on the port back
to Y, then it will just resurrected the entry in the
TimedExpirationMap and not write the new MAC to ZooKeeper.

This patch fixes this particular issue. Note that it's impossible to
order notifications from ZK with incoming packets on the bridge, so
there are still potential races that could happen if MACs are flapping
too quickly.

Change-Id: If489d746f4b7143faed70b5c1392551204f4177d
Ref: MNA-1012
Signed-off-by: Duarte Nunes <duarte@midokura.com>