VxGW: apply second and later bindings to OVSDB
Bug fix: a VtepSynchronizer that's watching a VTEP will *not* apply
bindings to a second and later physical port.

We're not updating the unit tests here to cover this case because our
OVSDB clients are mocked with mockito, and it's quite tricky to verify
the sequences of updates that are expected to a VTEP.  The mock OVSDB
client is still pending fixing some bugs, so it can't be used either.

We've verified the patch with MDTS tests that are already merged in
master, and will be merged shortly in this branch.

Ref.: MNA-1003
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: Iad3f183d25cedede5601e70789f93796cf31da16