VxGW: apply second and later bindings to OVSDB
Bug fix: a VtepSynchronizer that's watching a VTEP will *not* apply
bindings to a second and later physical port.

We're not updating the unit tests here to cover this case because our
OVSDB clients are mocked with mockito, and it's quite tricky to verify
the sequences of updates that are expected to a VTEP.  The mock OVSDB
client is still pending fixing some bugs, so it can't be used either.

Since we're adding MDTS tests for the VTEP that cover this case, we'll
take this as sufficient coverage for now.

Ref.: MNA-1003
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: I7b0813fc84173cd09f5d4080d3ed924a9be6a216