Fixing JAVA_LIBRARY_PATH for mn-conf
OpenJDK uses the system jnidispatcher as default, since /usr/lib/jni
is in the java library path. We specify our jna.jar in the classpath,
this leads to incompatibility. We should use either (1) the system
jnidispatcher and the system jna.jar or (2) the packaged jnidispatcher
and the packaged jna.jar. In this fix, we remove the /usr/lib/jni from
the library path to use the packaged jnidispatcher.

Ref: MNA-1017

Change-Id: I903f6df9ba0c5ea3bc3c255d67f249736b7a9c67
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>