Fixing JAVA_LIBRARY_PATH for mn-conf
OpenJDK uses the system jnidispatcher as default, since /usr/lib/jni
is in the java library path. We specify our jna.jar in the classpath,
this leads to incompatibility. We should use either (1) the system
jnidispatcher and the system jna.jar or (2) the packaged jnidispatcher
and the packaged jna.jar. In this fix, we remove the /usr/lib/jni from
the library path to use the packaged jnidispatcher.

Ref: MNA-1017

Change-Id: I457e7e8d015997958282cf18f0afcb4f55549d89
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>
(cherry picked from commit fc99e8a1dddc4a61fb91dc399804e2919453aaf2)