Update patch set 2
Patch Set 2:

This patch is a clone of https://review.gerrithub.io/#/c/250495/ that does NOT rely on the MDTS patch adding vxgw tests.

This is because there is a failure in the tests (that doesn't happen in master) that is unrelated to this patch.

The fix has been verified manually, and the test can be seen passing on the patch referenced above, so we're reasonably confident of this patch. We'll merge this first so that we can unblock RC4, and merge the MDTS test later. (Waiting to merge the code changes until the test is in is not feasible as it'd invalidate the RC4 QA).

Patch-set: 2