Add script to query CI targets
This patch provides a script that allows a developer (or the CI system)
get the right gradle tasks and flags that should be used to reliably
verify a given git ref.  This patch only considers v5 code (e.g., it
won't give the right targets for v1 code or the period when there was
dual stack).

Typically, usage will be:

    ./what_ci_targets.sh

Which will assume HEAD.  A different git ref can be passed as first
parameter otherwise:

    ./what_ci_targets.sh 7365846

Change-Id: Ib1c258a14512352e6673417c7d334bcf99cf2f91
Signed-off-by: Galo Navarro <galo@midokura.com>