MDTS: increase timeout to wait for default route
BGP tests fails intermittently when the test waits for the default route
to be announced to midolman peers. Let's increase the timeout to
overcome this issue in slow or overloaded platforms. We also bump the
midonet-sandbox revision which contains changes to the images to
increase the timeout when waiting for an external interface to be set
up which made bgp tests flaky. A couple more usages of "raise e" removed
to preserve the original exception context.

Ref.: MNA-1023

Change-Id: I841c4772c6c219edba0bc91043b1fbbc651edc09
Signed-off-by: Xavi León <xavi.leon@midokura.com>