C3PO: Set device ID on router interface create
On router interface create, when a port ID is specified instead of
subnet ID, it is possible that the port was not previously a router
interface.  In that case, device ID must be updated with the router ID
as it may be needed in other places in C3PO translations.  Device owner
was already correctly updated.

Change-Id: Ic034f4289ba5cb1ccc87745ace9e7bac884c3ce7
Fixes: MNA-1028
(cherry picked from commit 907c64d0235508e0dff911ee600464ceeb54bebc)