CI recommendations: fix bug and enhance
This patch fixes a bug in the what_ci_targets.sh script that wasn't
applying the MDTS count when chosing whether to recommend a full MDTS
build.

It additionaly fixes and enhances the checks made to detect whether the
change is just a version bump.

Also, add a :tempest tag.

Change-Id: Idb80990fc090c4b4631e650bea5ec214c6efc44f
Signed-off-by: Galo Navarro <galo@midokura.com>