VxGW: avoid eager initialization of watchers
Even if the VxGW is set to disabled in config, subresources are actually
ran on every node.

This is because the service is configured as an eager singleton in
Guice, which will instantiate the VxlanGatewayService *before* any kind
of initialization conditional to config.  Said class initializes upon
construction a couple of ZK monitoring classes for hosts and tunnel
zones, whose constructors start to watch ZK immediately.  This results
in monitors being set in ZK, even if later the API bootstrap decides not
to start the VxGW service.

This patch moves these ZK watcher properties in the VxlanGatewayService
class lazily initialized, only when the service is actually started
(doStart as per the Guice workflow.)

Ref.: MNA-1033
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: Ieb56e69d49d7f39911a4bb07779d33be06cbf634