VxGW: NPE if tz member is deleted
This patch prevents a NPE while calculating the flooding proxy for a
tunnel zone when one of its members is deleted (from ZK) while
processing.

This bug was found due to a data inconsistency in an environment that
manipulated ZK, removing a host but not the tunnel zone membership.
This resulted in the TunnelZoneState going to fetch a host, receiving a
null, and choking on it.

Ref.: MNA-1032
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: I8300b81f1951f675516a24ba854dfa8bc879451a