API: Fixes host update operation
This patch fixes the following known issues:

- Host DTOs tracks the back-references to the host tunnel-zones to
ensure that when a host is updated (the only supported update is
changing the flooding proxy weight), the tunnel-zone membership is
not lost.

- When a host is deleted, its membership in tunnel-zones is removed
atomically for those fields that are not protected by ZOOMs
referential integrity. This ensures that a deleted host does not
appear in the membership list of a tunnel-zones.

- Includes error messages in the API error responses.

Ref: MNA-1031

Change-Id: I402b260db9127ae4441dd95dda0c3e5b4f78569c
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
Signed-off-by: Galo Navarro <galo@midokura.com>