Agent: Remove from waiting room in case of error
When a simulation was postponed and the future terminated with an
error, we were not removing the packet context from the waiting room.
The context would later be removed when dealing with expired entries,
but this would emit a warning. This patch fixes this so no misleading
warning is emitted.

Change-Id: Ib738cf1b87a17a41de0b9fc8bc22744d580c566f
Ref: MN-4030
Signed-off-by: Duarte Nunes <duarte@midokura.com>