CLI: fix embedded lists
The lambda assigned to the _dto attribute in objects created
for elements of embedded collections were all pointing to the same
element. Reason: python lambdas capture scope lexically (!?!?).
This patch fixes the lambda definition to actually return the
element it's supposed to, by adding a parameter to the lambda
and assigning the value we want to close on as its default value.

Change-Id: I848e3fa4b16400788c2b808be1c8b460be5c1c5a
Fixes: MNA-1004
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>