Update patch set 2
Patch Set 2:

Your call, but I think it would be better to leave the console appender definition in, but just not include it in the root logger. That makes it easier to add back in when needed.

Maybe not in staging/v5.0, but we should leave it there in master.

Patch-set: 2
Reviewer: bberg@midokura.com <1002821@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0