API: Application DTO automatically includes resources
This ensures that the field of the Application DTO are computed
from the resources currently loaded by the API. Each resource may
use the ApiResource annotation to specify a name and a template,
which are used as the field names. Both are optional, and if not
present the Application DTO will not include that resource.

The patch also removes support for the `host_version` and
`write_version` end-points which were lacking implementation in
the MidoNet v5.0 API.

Change-Id: Ibc2eb6abe8690d8abc2a98e0b2b9094912d11960
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
(cherry picked from commit 4f2effea423fa67f0e2f179ce21204e558d2b759)