Update patch set 2
Patch Set 2:

(1 comment)

Since this is a bug fix that should be merged soon, I simplified the retry policy by simply sending a message to the actor to retry the write operation (no sleep take place anymore). I also think that if we want to centralize the retry mechanism this should be discussed more broadly.

Patch-set: 2
Label: Code-Review=0