Prevent ConcurrentModificationException in test
Found it in CI.  There are 2 subscribers to a given ZK node.  An
eventually block is capturing notifications on one of the watchers
waiting for the last element to be emitted.  Immediately after, we check
on the second subscriber to verify that we got the same value.

We hit the CME because the notification thread happened to notify
exactly when the events.last was being done.  Events is a java list, and
the .last hides a surprise as it is actually implemented with an
interator behind the scenes:

def last: A = {
    var lst = head
    for (x <- this)
        lst = x
    lst
}

This, by the way, is a very likely cause of flakiness on this test which
was pending investigation: should the notification have arrived an
instant later than the check, the test would've failed.

Change-Id: I48d7b3d6c857dd8dcf3ec1d88af87985acfecc20
Signed-off-by: Galo Navarro <galo@midokura.com>
(cherry picked from commit b6db55223a90d8d9aeec5611a8b4f9a7c296c850)