MDTS: Add keystone to the default_v2 flavour
Add support for keystone in the default_v2 flavour used on the gates.
Cluster related configuration to use keystone as auth provider is done
in the sandbox image and applied if the cluster service contains a link
to the keystone service in the flavours file. If the cluster service
is not linked to a keystone service it will default to use the MockAuth
auth provider.

Change-Id: I280ad76fabe76fa727866f94f655bce2cb61f623
Signed-off-by: Xavi León <xavi.leon@midokura.com>