Service insertion: Only mark outgoing packets as redirected
Service insertion ports should only accept traffic which is redirected
out of them. There's a flag in packet context for this. However if this
is set on all redirects, then broadcast traffic ends up going out the
service port, which may bounce it back and end up messing up the arp
table of the bridge. This change makes it so only packets which are
going to be redirected out of a port have redirect set, so service ports
will be able to drop broadcast packets.

Ref MNA-1046

Change-Id: I5da652f6601ccb4d4adde791c38c85d3b8204c29
Signed-off-by: Ivan Kelly <ivan@midokura.com>