MDTS: First Neutron based test (PoC)
This patch adds an initial test generating the topology with the Neutron
API. Using the neutron API, we create a net, a subnet and two ports.
This two ports are bound to two VMs located on different computes. Send
a ping using the same framework used previsouly for testing through the
internal midonet api.

TODO:
- move the test to a neutron directory for better organization.
- move run_tests.sh (and associated files) to a common parent.

Change-Id: I240cb8a0cc7a4d37d4912fc4f25dd1834cbb7ecb
Signed-off-by: Xavi León <xavi.leon@midokura.com>