Update patch set 5
Patch Set 5:

I think that's a worse design, to be accessing static stuff and making it more complex to test and mock. Also, we'd either have to add caching to the HostIdGenerator (if this.hostId eq null then check this and that) or cache the host id locally every place the HostIdProvider is used. With the HostIdProvider in the patch we at least get great inlining without the need to branch. (It's accessed per packet for flow state replication.)

Patch-set: 5