Agent: Further minimize reliance on Guice
This patch further minimizes the reliance on Guice from the Agent,
where all objects are virtually singletons. We now pass an Injector
into the MidolmanModule, where we can bootstrap everything. This will
be even more important in future patches, where it's there where we'll
bootstrap the packet processing threads. The alternative approach
would be to have a factory object into which we would inject
dependencies, but this creates unneeded complexity.

As of this patch, Guice is only needed in the Agent to bootstrap the
Guava services and the Akka actors.

Change-Id: Ib15f073ffab79a30d280b9f2f1c5ddd5d3287462
Ref: MNA-89
Signed-off-by: Duarte Nunes <duarte@midokura.com>