Clear context redirect flag after redirecting
Service ports only emit packets where the redirect flag has been set
on the context. This flag is set by the redirect chain rule action.

However, in the case of a flood, the same context will be used when
emitting to multiple ports. If one port ends up redirecting (because it
is protected by an insertion), any subsequent service port on the bridge
will see the redirect flag and emit also, which is incorrect.

This change fixes this by clearing the flag after emission.

Change-Id: Icbe38e18839123e3bb6f186f33a3ceddd1bcbdf2
Signed-off-by: Ivan Kelly <ivan@midokura.com>