ZOOM: get and getAll go to ZK always
Both methods were doing lookups in the local cache trying to fetch
objects requested to get and getAll.  This optimization is fine for
clients doing reads (e.g., the agent) but not for those doing both reads
and writes (e.g., the API) as a fast sequence of WRITE/GET operations
may yield an out of date version of the object.  This was detected in
new MDTS VxLAN Gateway tests, as tracked in the ref'd issue.

This patch removes the optimisation, having both methods go to ZK in all
cases for GETs.  The storage API is already async, so clients are
already forced to assume that we may need IO to retrieve the object.

The optimisation could still be valid, but should be sliced out to a
separate API method (or, implemented only in ReadonlyStorage).

Ref.: MNA-1041
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: I3374daea842d9a1b44c551d507a47d83f9a438d0