Update patch set 1
Patch Set 1: Code-Review-1

Pino is telling me we can avoid writing REST API and client for this as the APIs are being developed on the neutron side.  We have now support for neutron in MDTS so this simplifies the func. testing story, and translations are in-tree, which simplifies the unit testing story. What do you think?

Patch-set: 1
Reviewer: Galo Navarro <1002731@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1